Fixed inline initialization of Redactor.
When you 'Add another Foo' in inline forms in the admin, the Redactor editor
fields were breaking, causing mayhem. Data entry for Red Bull was lost (I was
never able to replicate it and thought it might be due to renaming tables) and
it was making Manero weekender email generation problematic.